Update table striping and hover modifiers
### Ditch bulk of table-row mixin

This whol mixin was confusing and troublesome. If you added the .table-* 
class to a row, it applied to the row and cells within—this caused 
translucent backgrounds to double up.

And now that border-color is inherited, there's also no need to be 
reseting and targeting that.

Also removed the .table-hover styles from the mixin because we're 
blending the hover and background-color through a little 
background-image gradient hack.

### Docs cleanup

- Moved striped to be after bordered examples
- Added explanation for how striped and hovered tables work now
- Added example for table-hover and background-colors via .table-* 
classes
- Added nesting example